+++ packages/react/src/internals/index.ts
@@ -1,4 +1,1 @@
-export { default as Section } from './Section';
-export { default as Header } from './Header';
-export * from './ListBoxItem';
-export { default as ListBox } from './ListBox';
+export * from './ListBox';
